Move migration tag tests to dataset tag test file
